extracted seeding, unlinked, and orphaned files services from DownloadCleaner
